requirements.txt
python-aternos==2.1.3
discord-py-interactions == 4.3.4
